ordorsystem/backend/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
Werkzeug>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
Jinja2>=3.0.0